+++ app/src/main/res/values/strings.xml
@@ -1,13 +1,4 @@
 <resources>
     <string name="app_name">Punch-AOS</string>
     <string name="title_activity_sample">SampleActivity</string>
-
-    <string name="matching_card_hint">친구 코드를 입력하고 매칭하기</string>
-    <string name="matching_card_title">"우리는 잘 맞을까?</string>
-    <string name="matching_card_caption">궁합, 공통점 등 친구의 새로운 정보를 알 수 있어요!</string>
-    <string name="my_code_card_caption">나의 코드</string>
-    <string name="my_code">%s</string>
-    <string name="my_profile_card_caption">내 프로필</string>
-    <string name="profile_view_counter_caption">내 프로필을</string>
-    <string name="profile_view_counter_card_subtitle">%d명이 조회했어요</string>
 </resources>
