+++ build-logic/convention/src/main/kotlin/com/moya/funch/plugins/AndroidHiltPlugin.kt
@@ -14,6 +14,7 @@ class AndroidHiltPlugin : Plugin<Project> {
             }
 
             dependencies {
+                add("implementation", libs.findLibrary("hilt.navigation.compose").get())
                 add("implementation", libs.findLibrary("hilt.android").get())
                 add("ksp", libs.findLibrary("hilt.compiler").get())
             }
